chore(js): rename berty-i18n to i18n
