+++ .github/workflows/integration.yml
@@ -16,7 +16,7 @@ on:
       - 'go.*'
       - '**.go'
       - 'js/**'
-      - '!js/packages/berty-i18n/locale/*/*.json'
+      - '!js/packages/i18n/locale/*/*.json'
       - '.github/workflows/integration.yml'
   pull_request:
     paths:
@@ -25,7 +25,7 @@ on:
       - 'go.*'
       - '**.go'
       - 'js/**'
-      - '!js/packages/berty-i18n/locale/*/*.json'
+      - '!js/packages/i18n/locale/*/*.json'
       - '.github/workflows/integration.yml'
 
 # FIXME:
