+++ .github/workflows/protobuf.yml
@@ -13,7 +13,7 @@ on:
       - "config/**"
       - "docs/Makefile"
       - ".github/workflows/protobuf.yml"
-      - "js/packages/berty-i18n/locale/en/messages.json"
+      - "js/packages/i18n/locale/en/messages.json"
       - "tool/generate-i18n-testlang"
       - "**/gen.sum"
       - "**.pb.go"
@@ -43,7 +43,7 @@ on:
       - "config/**"
       - "docs/Makefile"
       - ".github/workflows/protobuf.yml"
-      - "js/packages/berty-i18n/locale/en/messages.json"
+      - "js/packages/i18n/locale/en/messages.json"
       - "tool/generate-i18n-testlang"
       - "**/gen.sum"
       - "**.pb.go"
