+++ js/packages/i18n/locale/fr-FR/messages.json
