+++ js/packages/i18n/locale/ja-JP/messages.json
