+++ js/packages/i18n/locale/ko-KR/messages.json
