+++ js/packages/i18n/locale/no-NO/messages.json
