+++ js/packages/i18n/locale/pt-BR/messages.json
