+++ js/packages/i18n/locale/te-IN/messages.json
